pyproject.toml
requires = ["setuptools~=63.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["woob", "woob.*", "weboob"]

[project]
name = "woob"
version = "3.5"
description = "Woob, Web Outside Of Browsers"
authors = [
    { name="Romain Bignon", email="romain@woob.dev" }
]
readme = "README.rst"
license = { text="LGPL-3.0-or-later" }
keywords = ["scraping", "web", "banking"]
classifiers = [
    "Environment :: Console",
    "License :: OSI Approved :: GNU Lesser General Public License v3 or later (LGPLv3+)",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Browsers",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Text Processing :: Markup :: HTML"
]
requires-python = ">=3.7"
dependencies = [
    "lxml",
    "requests >= 2.0.0",
    "python-dateutil",
    "PyYAML",
    "html2text >= 3.200",
    "unidecode",
    "Pillow",
    "Babel",
    "packaging ~= 23.0",
    "pycountry",
]

[project.urls]
"Homepage" = "https://woob.tech"
"Release notes" = "https://gitlab.com/woob/woob/-/releases"
"Documentation" = "https://woob.dev"
"Bug Tracker" = "https://gitlab.com/woob/woob/-/issues"

[project.scripts]
woob = "woob.applications.main:WoobMain.run"

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "$version"
update_changelog_on_bump = true
version = "3.5"
version_files = [
    "pyproject.toml:version",
]

[tool.isort]
lines_after_imports = 2
include_trailing_comma = true
use_parentheses = true
no_inline_sort = true
multi_line_output = 5
known_first_party = ["woob", "woob_modules"]
line_length = 120

[tool.mypy]
show_error_codes = true

[tool.flake8]
max-line-length = 120
exclude = ['dist', '*.egg-info', 'build', '.git', '__pycache__', 'boilerplate_data']
extend-ignore =[
    # Disabled because it doesn't allow exceptions, for example URLs or log
    # messages shouldn't be split, less readable or searchable.
    'E501',
    # newline after binary operator
    'W503',
    # Too many leading '#' for block comment
    # But it's a nice visual separator sometimes.
    'E266',
    # flake8-future-import: postponed annotations
    'FI58',
    # do not perform function calls in argument defaults: breaks fastapi qs definitions
    'B008'
]


[tool.pytest.ini_options]
python_files = "tests/*.py test.py"
testpaths = [
    "tests",
    "woob/browser/browsers.py",
    "woob/browser/pages.py",
    "woob/browser/filters/standard.py",
    "woob/browser/filters/json.py",
    "woob/tools/json.py",
    "woob/tools/misc.py"
]
